Moving category filtering code out of main.ts
